+++ examples/sycl/06_pvc_flash_attention/06_pvc_decode_attention.cpp
@@ -29,7 +29,7 @@
  *
  **************************************************************************************************/
 
-#include "pvc_flash_decode_runner.hpp"
+#include "pvc_flash_attn_decode_runner.hpp"
 
 int main(int argc, const char **argv) {
   //
