+++ examples/sycl/06_pvc_flash_attention/CMakeLists.txt
