+++ examples/sycl/10_pvc_flash_attention_cachedKV/10_pvc_prefill_attention_prefill_cachedKV.cpp
@@ -51,7 +51,7 @@
     Call with `--help` for information about available options
 */
 
-#include "pvc_flash_attn_cachedKV_runner.hpp"
+#include "pvc_flash_attn_prefill_cachedKV_runner.hpp"
 
 int main(int argc, const char **argv) {
   //
