+++ test/unit/flash_attention/flash_attention_decode/flash_decode_testbed_3x.hpp
@@ -39,7 +39,7 @@
 #include "flash_attention_v2/collective/fmha_fusion.hpp"
 #include "flash_attention_v2/kernel/tile_scheduler.hpp"
 #include "cutlass/util/packed_stride.hpp"
-#include "flash_attention_v2/kernel/xe_flash_attn_decode_gemm.hpp"
+#include "flash_attention_v2/kernel/xe_flash_attn_decode.hpp"
 #include "flash_attention_v2/collective/xe_flash_attn_decode_epilogue.hpp"
 #include "flash_attention_v2/collective/xe_flash_attn_decode_softmax_epilogue.hpp"
 #include "cutlass/util/GPU_Clock.hpp"
